Feat: Add Game Table with Seeds, and newgame endpoint

+++ server/db/migrations/20221101080109_create_playlist_table.js
@@ -17,6 +17,12 @@ exports.up = function(knex) {
  * @param { import("knex").Knex } knex
  * @returns { Promise<void> }
  */
-exports.down = function(knex) {
-  return knex.schema.dropTable('playlist_data');
+exports.down = async function(knex) {
+  await knex.schema
+  .hasTable('playlist_data')
+  .then(async function(exists) {
+    if (exists) {
+        return knex.schema.dropTable('playlist_data');
+      };
+  });
 };
